Fix implicit numeric widening warnings

+++ axle-core/src/main/scala/axle/awt/Image.scala
@@ -54,11 +54,11 @@ object Image {
         val lastRow = height - 1
         val image = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB)
         (0 until width) foreach { column =>
-          val x0 = scaledArea.unframeX(column)
-          val x1 = scaledArea.unframeX(column + 1)
+          val x0 = scaledArea.unframeX(column.toDouble)
+          val x1 = scaledArea.unframeX(column + 1d)
           (0 until height) foreach { row =>
-            val y0 = scaledArea.unframeY(row)
-            val y1 = scaledArea.unframeY(row + 1)
+            val y0 = scaledArea.unframeY(row.toDouble)
+            val y1 = scaledArea.unframeY(row + 1d)
             val rgb = Color.toRGB(c(f(x0, x1, y0, y1)))
             image.setRGB(column, lastRow - row, rgb)
           }
